+++ pkg/query-service/rules/setups_test.go
@@ -40,7 +40,12 @@ func prepareQuerierForMetrics(t *testing.T, telemetryStore telemetrystore.Teleme
 		telemetryStore,
 		metadataStore,
 		nil, // prometheus
-		&statementbuilder.Builders{Metric: metricStmtBuilder},
+		nil, // traceStmtBuilder
+		nil, // logStmtBuilder
+		nil, // auditStmtBuilder
+		metricStmtBuilder,
+		nil, // meterStmtBuilder
+		nil, // traceOperatorStmtBuilder
 		nil, // bucketCache
 		fl,
 		0,
@@ -69,7 +74,12 @@ func prepareQuerierForLogs(t *testing.T, telemetryStore telemetrystore.Telemetry
 		telemetryStore,
 		metadataStore,
 		nil, // prometheus
-		&statementbuilder.Builders{Log: logStmtBuilder},
+		nil, // traceStmtBuilder
+		logStmtBuilder,
+		nil, // auditStmtBuilder
+		nil, // metricStmtBuilder
+		nil, // meterStmtBuilder
+		nil, // traceOperatorStmtBuilder
 		nil, // bucketCache
 		fl,
 		5*time.Minute, // logTraceIDWindowPadding
@@ -99,7 +109,12 @@ func prepareQuerierForTraces(t *testing.T, telemetryStore telemetrystore.Telemet
 		telemetryStore,
 		metadataStore,
 		nil, // prometheus
-		&statementbuilder.Builders{Trace: traceStmtBuilder},
+		traceStmtBuilder,
+		nil, // logStmtBuilder
+		nil, // auditStmtBuilder
+		nil, // metricStmtBuilder
+		nil, // meterStmtBuilder
+		nil, // traceOperatorStmtBuilder
 		nil, // bucketCache
 		fl,
 		0,
