+++ pkg/statementbuilder/tracesstatementbuilder/statement_builder.go
@@ -10,6 +10,7 @@ import (
 	"github.com/SigNoz/signoz/pkg/factory"
 	"github.com/SigNoz/signoz/pkg/flagger"
 	"github.com/SigNoz/signoz/pkg/querybuilder"
+	"github.com/SigNoz/signoz/pkg/statementbuilder"
 	"github.com/SigNoz/signoz/pkg/statementbuilder/resourcefilter"
 	"github.com/SigNoz/signoz/pkg/telemetryschema/tracestelemetryschema"
 	"github.com/SigNoz/signoz/pkg/telemetrystore"
@@ -34,6 +35,28 @@ type traceQueryStatementBuilder struct {
 }
 
 var _ qbtypes.StatementBuilder[qbtypes.TraceAggregation] = (*traceQueryStatementBuilder)(nil)
+
+// NewFactory returns a provider factory for the trace query statement builder. Its
+// New internalizes the FieldMapper, ConditionBuilder, and AggExprRewriter, and reads
+// SkipResourceFingerprint from the config.
+func NewFactory(
+	telemetryStore telemetrystore.TelemetryStore,
+	metadataStore telemetrytypes.MetadataStore,
+	fl flagger.Flagger,
+) factory.ProviderFactory[qbtypes.StatementBuilder[qbtypes.TraceAggregation], statementbuilder.Config] {
+	return factory.NewProviderFactory(
+		factory.MustNewName("traces"),
+		func(_ context.Context, settings factory.ProviderSettings, cfg statementbuilder.Config) (qbtypes.StatementBuilder[qbtypes.TraceAggregation], error) {
+			fm := tracestelemetryschema.NewFieldMapper()
+			cb := tracestelemetryschema.NewConditionBuilder(fm)
+			aggExprRewriter := querybuilder.NewAggExprRewriter(settings, nil, fm, cb, fl)
+			return NewTraceQueryStatementBuilder(
+				settings, metadataStore, fm, cb, aggExprRewriter, telemetryStore, fl,
+				cfg.SkipResourceFingerprint.Enabled, cfg.SkipResourceFingerprint.Threshold,
+			), nil
+		},
+	)
+}
 
 func NewTraceQueryStatementBuilder(
 	settings factory.ProviderSettings,
